+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.1.0] - 2023-10-07
+
+### Changed
+- `exekias query ...` command is now `exekias runs query ...`.
+  It doesn't show hidden runs by default.
+  `exekias query ... --hidden` command shows only hidden runs instead.
+
+### Added
+- `exekias runs show <run>` command to print run metadata.
+- `exekias runs hide <run> [--unhide]` command to hide/unhide runs.
+
+
 ## [2.0.5] - 2023-09-26
 ## [2.0.4] - 2023-09-26